+++ config/test.exs
@@ -23,6 +23,7 @@ config :exshome,
     {ExshomePlayer.Services.MpvServer, ExshomeTest.Hooks.MpvServer},
     {Exshome.PubSub, ExshomeTest.Hooks.PubSub},
     {Exshome.Repo, ExshomeTest.Hooks.Repo},
+    {Exshome.Variable, ExshomeTest.Hooks.Variable},
     {:live_view, [ExshomeTest.Hooks.LiveView]}
   ]