+++ lib/exshome/dependency/gen_server_dependency.ex
@@ -90,7 +90,7 @@ defmodule Exshome.Dependency.GenServerDependency do
   end
 
   @spec dependency_key(Dependency.dependency()) :: any()
-  def dependency_key(dependency), do: {__MODULE__, dependency}
+  defp dependency_key(dependency), do: {__MODULE__, dependency}
 
   @spec get_value(Dependency.dependency()) :: any()
   def get_value(dependency) do