+++ lib/exshome_automation/web/live/preview.ex
@@ -2,7 +2,8 @@ defmodule ExshomeAutomation.Web.Live.Preview do
   @moduledoc """
   Automation preview widget.
   """
+  alias ExshomeAutomation.Services.AutomationStatus
 
   use ExshomeWeb.Live.AppPage,
-    dependencies: []
+    dependencies: [{AutomationStatus, :status}]
 end